[Cidob] Migration Scripts

+++ Scripts/01_CREATE_FILL_Table_Arch.sql
@@ -1,4 +1,5 @@
-DROP TABLE dbo.Arch
+if (object_id('Arch') is not null)
+	DROP TABLE dbo.Arch
 GO
 CREATE TABLE [dbo].[Arch](
 	[Id] [int] IDENTITY(1,1) NOT NULL,
@@ -27,15 +28,15 @@ GO
 GO
 GO
 GO
-INSERT INTO dbo.Arch VALUES (1,2,'Rígido','|Arco|',1)
-INSERT INTO dbo.Arch VALUES (2,3,'Semirígido','|Semirígido|',1)
-INSERT INTO dbo.Arch VALUES (3,4,'Blando','|Blando|',1)
+INSERT INTO dbo.Arch VALUES (4,2,'Rígido','|Arco|',1)
+INSERT INTO dbo.Arch VALUES (4,3,'Semirígido','|Semirígido|',1)
+INSERT INTO dbo.Arch VALUES (4,4,'Blando','|Blando|',1)
 INSERT INTO dbo.Arch VALUES (4,5,'Rígido Alto','|Arco Alto|',1)
-INSERT INTO dbo.Arch VALUES (5,6,'Semirígido Alto','|Semiríg Alto|',1)
-INSERT INTO dbo.Arch VALUES (6,7,'Blando Alto','|Blando Alto|',1)
-INSERT INTO dbo.Arch VALUES (7,8,'Rígido Bajo','|Arco Bajo|',1)
-INSERT INTO dbo.Arch VALUES (8,9,'Semirígido Bajo','|Semiríg Bajo|',1)
-INSERT INTO dbo.Arch VALUES (9,10,'Blando Bajo','|Blando Bajo|',1)
+INSERT INTO dbo.Arch VALUES (4,6,'Semirígido Alto','|Semiríg Alto|',1)
+INSERT INTO dbo.Arch VALUES (4,7,'Blando Alto','|Blando Alto|',1)
+INSERT INTO dbo.Arch VALUES (4,8,'Rígido Bajo','|Arco Bajo|',1)
+INSERT INTO dbo.Arch VALUES (4,9,'Semirígido Bajo','|Semiríg Bajo|',1)
+INSERT INTO dbo.Arch VALUES (4,10,'Blando Bajo','|Blando Bajo|',1)
 GO
 GO
 GO
@@ -43,12 +44,12 @@ GO
 GO
 GO
 GO
-INSERT INTO dbo.Arch VALUES (17,11,'Apoyo Total','|Apoyo Total|',1)
-INSERT INTO dbo.Arch VALUES (18,12,'Apoyo Total (Bajo)','|A.Total (Bajo)|',1)
-INSERT INTO dbo.Arch VALUES (19,13,'Apoyo Total (Arco Debajo)','|A.Total|(Arc.Abajo)|',1)
+INSERT INTO dbo.Arch VALUES (4,11,'Apoyo Total','|Apoyo Total|',1)
+INSERT INTO dbo.Arch VALUES (4,12,'Apoyo Total (Bajo)','|A.Total (Bajo)|',1)
+INSERT INTO dbo.Arch VALUES (4,13,'Apoyo Total (Arco Debajo)','|A.Total|(Arc.Abajo)|',1)
 GO
-INSERT INTO dbo.Arch VALUES (35,14,'Semiblando','|Semiblando|',1)
-INSERT INTO dbo.Arch VALUES (36,15,'1/4 de esfera','|1/4 de esfera|',1)
-INSERT INTO dbo.Arch VALUES (37,16,'Asimetrico Rígido','|Asim. Rígido|',1)
-INSERT INTO dbo.Arch VALUES (38,17,'Arco Rigido Recto 90º Casa Cuna','|Rigido 90º|',1)
-INSERT INTO dbo.Arch VALUES (39,18,'Arco Blando Recto 90º Casa Cuna','|Blando 90º|',1)
+INSERT INTO dbo.Arch VALUES (4,14,'Semiblando','|Semiblando|',1)
+INSERT INTO dbo.Arch VALUES (4,15,'1/4 de esfera','|1/4 de esfera|',1)
+INSERT INTO dbo.Arch VALUES (4,16,'Asimetrico Rígido','|Asim. Rígido|',1)
+INSERT INTO dbo.Arch VALUES (4,17,'Arco Rigido Recto 90º Casa Cuna','|Rigido 90º|',1)
+INSERT INTO dbo.Arch VALUES (4,18,'Arco Blando Recto 90º Casa Cuna','|Blando 90º|',1)
